feat(offline): split offline related data models to separate files
