+++ src/index.ts
@@ -10,6 +10,4 @@ export * from './source';
 export * from './subtitleTrack';
 export * from './styleConfig';
 export * from './ui';
-export * from './offline/offlineContentOptions';
-export * from './offline/offlineContentManager';
-export * from './offline/offlineContentManagerListener';
+export * from './offline';
