+++ src/player.ts
@@ -7,8 +7,7 @@ import { AudioTrack } from './audioTrack';
 import { SubtitleTrack } from './subtitleTrack';
 import { StyleConfig } from './styleConfig';
 import { TweaksConfig } from './tweaksConfig';
-import { OfflineContentManager } from './offline/offlineContentManager';
-import { OfflineSourceOptions } from './offline/offlineContentOptions';
+import { OfflineContentManager, OfflineSourceOptions } from './offline';
 
 const PlayerModule = NativeModules.PlayerModule;
